requirements.txt
langchain==0.1.12
langchain-aws==0.1.6
chromadb==0.4.24
torch==2.2.1
transformers==4.38.2
sentence-transformers==2.5.1
boto3==1.34.69
streamlit==1.32.2
python-dotenv==1.0.1
pypdf==4.0.2